requirements.txt
<<<<<<< HEAD
adium-theme-ubuntu==0.3.4
autopep8==1.3.3
Django==1.11.7
pycodestyle==2.3.1
pycurl==7.43.0
pygobject==3.20.0
python-apt==1.1.0b1
pytz==2017.3
unity-lens-photos==1.0
youtube-dl==2016.2.22
=======
Babel==2.5.1
confusable-homoglyphs==2.0.2
Django==2.0.1
django-bootstrap==0.2.4
django-bootstrap3==9.1.0
django-phonenumber-field==2.0.0
django-registration==2.3
phonenumberslite==8.8.9
Pillow==5.0.0
psycopg2==2.7.3.2
pytz==2017.3
>>>>>>> 9fff37582771a5b9fbfd826b8060adfbdcd59563
